+++ clients/admin-ui/src/features/taxonomy/DataCategoryChecklist.tsx
@@ -4,14 +4,14 @@ import { useMemo, useState } from "react";
 import { DataCategory } from "~/types/api";
 
 import CheckboxTree from "../common/CheckboxTree";
-import { transformDataCategoriesToNodes } from "./helpers";
+import { transformTaxonomyEntityToNodes } from "./helpers";
 
 interface Props {
   dataCategories: DataCategory[];
 }
 const DataCategoryChecklist = ({ dataCategories }: Props) => {
   const nodes = useMemo(
-    () => transformDataCategoriesToNodes(dataCategories),
+    () => transformTaxonomyEntityToNodes(dataCategories),
     [dataCategories]
   );
   const [checked, setChecked] = useState<string[]>([]);