+++ tests/ctl/api/test_validate.py
@@ -1,5 +1,6 @@
 # pylint: disable=missing-docstring, redefined-outer-name
-from json import dumps
+from base64 import b64decode
+from json import dumps, loads
 from os import getenv
 
 import pytest
@@ -15,6 +16,12 @@ EXTERNAL_CONFIG_BODY = {
         "aws_access_key_id": getenv("AWS_ACCESS_KEY_ID", ""),
         "aws_secret_access_key": getenv("AWS_SECRET_ACCESS_KEY", ""),
     },
+    "bigquery": {
+        "dataset": "fidesopstest",
+        "keyfile_creds": loads(
+            b64decode(getenv("BIGQUERY_CONFIG", "e30=").encode("utf-8")).decode("utf-8")
+        ),
+    },
     "okta": {
         "orgUrl": "https://dev-78908748.okta.com",
         "token": getenv("OKTA_CLIENT_TOKEN", ""),
@@ -23,7 +30,7 @@ EXTERNAL_CONFIG_BODY = {
 
 
 @pytest.mark.external
-@pytest.mark.parametrize("validate_target", ["aws", "okta"])
+@pytest.mark.parametrize("validate_target", ["aws", "okta", "bigquery"])
 def test_validate_success(
     test_config: FidesctlConfig,
     validate_target: str,
@@ -57,16 +64,27 @@ EXTERNAL_FAILURE_CONFIG_BODY = {
         "orgUrl": "https://dev-78908748.okta.com",
         "token": "INVALID_TOKEN",
     },
+    "bigquery": {
+        "dataset": "fidesopstest",
+        "keyfile_creds": loads(
+            b64decode(getenv("BIGQUERY_CONFIG", "e30=").encode("utf-8")).decode("utf-8")
+        ),
+    },
 }
+
+EXTERNAL_FAILURE_CONFIG_BODY["bigquery"]["keyfile_creds"][
+    "project_id"
+] = "INVALID_PROJECT_ID"
 
 EXPECTED_FAILURE_MESSAGES = {
     "aws": "Authentication failed validating config. The security token included in the request is invalid.",
     "okta": "Authentication failed validating config. Invalid token provided",
+    "bigquery": "Unexpected failure validating config. Invalid project ID 'INVALID_PROJECT_ID'. Project IDs must contain 6-63 lowercase letters, digits, or dashes. Some project IDs also include domain name separated by a colon. IDs must start with a letter and may not end with a dash.",
 }
 
 
 @pytest.mark.external
-@pytest.mark.parametrize("validate_target", ["aws", "okta"])
+@pytest.mark.parametrize("validate_target", ["aws", "okta", "bigquery"])
 def test_validate_failure(
     test_config: FidesctlConfig,
     validate_target: str,