+++ tests/ctl/cli/test_cli.py
@@ -2,7 +2,6 @@
 import os
 from base64 import b64decode
 from json import dump, loads
-from pathlib import PosixPath
 from typing import Generator
 
 import pytest
@@ -13,6 +12,13 @@ from py._path.local import LocalPath
 from fidesctl.cli import cli
 
 OKTA_URL = "https://dev-78908748.okta.com"
+
+
+def git_reset(change_dir: str) -> None:
+    """This fixture is used to reset the repo files to HEAD."""
+
+    git_session = Repo().git()
+    git_session.checkout("HEAD", change_dir)
 
 
 @pytest.fixture()
@@ -101,21 +107,50 @@ def test_dry_diff_apply(test_config_path: str, test_cli_runner: CliRunner) -> No
 
 
 @pytest.mark.integration
-def test_sync(
-    test_config_path: str, test_cli_runner: CliRunner, tmp_path: PosixPath
-) -> None:
-    """
-    Due to the fact that this command checks the real git status, a pytest
-    tmp_dir can't be used. Consequently a real directory must be tested against
-    and then reset.
-    """
-    test_dir = "demo_resources/"
-    result = test_cli_runner.invoke(cli, ["-f", test_config_path, "sync", test_dir])
-    print(result.output)
-    assert result.exit_code == 0
+class TestPull:
+    def test_pull(
+        self,
+        test_config_path: str,
+        test_cli_runner: CliRunner,
+    ) -> None:
+        """
+        Due to the fact that this command checks the real git status, a pytest
+        tmp_dir can't be used. Consequently a real directory must be tested against
+        and then reset.
+        """
+        test_dir = ".fides/"
+        result = test_cli_runner.invoke(cli, ["-f", test_config_path, "pull", test_dir])
+        git_reset(test_dir)
+        print(result.output)
+        assert result.exit_code == 0
 
-    git_session = Repo().git()
-    git_session.checkout("HEAD", test_dir)
+    def test_pull_all(
+        self,
+        test_config_path: str,
+        test_cli_runner: CliRunner,
+    ) -> None:
+        """
+        Due to the fact that this command checks the real git status, a pytest
+        tmp_dir can't be used. Consequently a real directory must be tested against
+        and then reset.
+        """
+        test_dir = ".fides/"
+        test_file = ".fides/test_resources.yml"
+        result = test_cli_runner.invoke(
+            cli,
+            [
+                "-f",
+                test_config_path,
+                "pull",
+                test_dir,
+                "-a",
+                ".fides/test_resources.yml",
+            ],
+        )
+        git_reset(test_dir)
+        os.remove(test_file)
+        print(result.output)
+        assert result.exit_code == 0
 
 
 @pytest.mark.integration